Add spec for serialize/deserialize DataQuery
